LayoutTests/http/tests/inspector/dom/disconnect-dom-tree-after-main-frame-navigation-expected.txt

Testing that DOMTrees are not retained by DOMTreeManager across main frame navigations.



== Running test suite: DOM.DisconnectTreeOnMainResourceChange
-- Running test case: CheckLazyInitializationOfDOMTree
PASS: There should not be a DOMTree listening to DOMTreeManager events initially.
DOMTree instance count: 0

-- Running test case: CheckDOMTreeCountAfterUsingGetter
PASS: There should be a one DOMTree listening to DOMTreeManager events after creation.
DOMTree instance count: 1
PASS: There should be four DOMTrees listening to DOMTreeManager events after touching each Frame.
DOMTree instance count: 4

-- Running test case: CheckDOMTreeCountAfterReloading
PASS: There should not be any DOMTrees listening to DOMTreeManager events after a main frame navigation.
DOMTree instance count: 0
